change rule_* to roll_* for nomenclature purposes; rename rules_usecase to check_usecase and move some functions to another usecase files
