+++ usecase/rule/check_usecase_test.go
@@ -1,0 +1,13 @@
+package rule
+
+import (
+	"testing"
+)
+
+func TestCheckPreparedSpellsList(t *testing.T) {}
+
+func TestCheckCantripsKnownList(t *testing.T) {}
+
+func TestCheckKnownList(t *testing.T) {}
+
+func TestCheckArmorClass(t *testing.T) {}
