+++ usecase/rule/roll_attack_usecase.go
