+++ usecase/rule/roll_attack_usecase_test.go
