+++ usecase/rule/roll_character_usecase.go
