+++ usecase/rule/roll_character_usecase_test.go
