+++ usecase/rule/roll_checks_usecase_test.go
