+++ usecase/rule/roll_features_usecase.go
@@ -9,220 +9,6 @@ import (
 	"github.com/betorvs/playbypost-dnd/utils"
 )
 
-//ListInformation func
-func ListInformation(name, value string) (*rule.SimpleList, error) {
-	list := new(rule.SimpleList)
-	switch name {
-	case "race", "races":
-		list.List = RaceList()
-		if value == "subrace" {
-			list.List = RaceListWithSubrace()
-		}
-
-	case "class", "classes":
-		list.List = ClassList()
-
-	case "background", "backgrounds":
-		list.List = BackgroundList()
-
-	case "ability":
-		list.List = AbilityList()
-
-	case "alignment":
-		list.List = AlignmentList()
-
-	case "skills", "skill":
-		list.List = SkillListAbility()
-
-	case "subrace", "subraces":
-		list.List = SubraceList(value)
-
-	case "condition", "conditions":
-		list.List = ListConditions()
-
-	case "damage", "damagetype":
-		list.List = ListOfDamageTypes()
-	}
-	return list, nil
-}
-
-//FullDescription func
-func FullDescription(kind, name, subname string) (*rule.FullDescription, error) {
-	fullDescription := new(rule.FullDescription)
-	switch kind {
-	case "race":
-		fullDescription.Description = RaceTraits(name, subname)
-	case "class":
-		fullDescription.Description = ClassInfo(name)
-
-	case "background":
-		m, _ := BackgroundStatistics(name)
-		fullDescription.Description = m
-
-	case "condition", "conditions":
-		d := conditionsMap(name)
-		fullDescription.Description = d
-
-	case "damage", "damagetype":
-		d := damageTypeMap(name)
-		fullDescription.Description = d
-	}
-
-	return fullDescription, nil
-}
-
-//GetConditions func
-func GetConditions(condition string, level int) *rule.ReturnCondition {
-	message := new(rule.ReturnCondition)
-	desc, dis, auto := conditionsCheck(condition, level)
-	message.Name = condition
-	message.Description = desc
-	message.Disvantages = dis
-	message.AutoFail = auto
-	return message
-}
-
-//CheckPreparedSpellsList func
-func CheckPreparedSpellsList(spellList *rule.PreparedSpellsList) (*rule.PreparedSpellsList, error) {
-	ability := AbilityForSpell(spellList.Class)
-	abilityModifier := CalcAbilityModifier(spellList.Ability[ability])
-	numberPreparedSpell := spellList.Level + abilityModifier
-	if len(spellList.PreparedSpells) != numberPreparedSpell {
-		err := fmt.Errorf("invalid number of prepared spells")
-		return spellList, err
-	}
-	for _, s := range spellList.PreparedSpells {
-		if !utils.StringInSlice(s, GetFullSpellList(spellList.Class)) {
-			err := fmt.Errorf("prepared spell %s is not in spell list for %s", s, spellList.Class)
-			return spellList, err
-		}
-	}
-	spellList.Verified = true
-	return spellList, nil
-}
-
-//CheckCantripsKnownList func
-func CheckCantripsKnownList(spellList *rule.KnownCantripList) (*rule.KnownCantripList, error) {
-	numberCantrips := spellList.CantripsKnown
-	if len(spellList.CantripsList) != numberCantrips {
-		err := fmt.Errorf("invalid number of cantrips")
-		return spellList, err
-	}
-	class := spellList.Class
-	if spellList.Class == "fighter" || spellList.Class == "rogue" {
-		if utils.StringInSlice("archetype-eldritch-knight-spellcasting", spellList.ClassFeatures) {
-			class = "eldritch-knight"
-		}
-		if utils.StringInSlice("archetype-arcane-trickster-spellcasting", spellList.ClassFeatures) {
-			class = "arcane-trickster"
-		}
-	}
-	cantrips := GetSpellListByClass(class, 0)
-	if utils.StringInSlice("domain-nature", spellList.ClassFeatures) {
-		druidCantrips := GetSpellListByClass(class, 0)
-		cantrips.List = append(cantrips.List, druidCantrips.List...)
-	}
-
-	for _, s := range spellList.CantripsList {
-		if !utils.StringInSlice(s, cantrips.List) {
-			err := fmt.Errorf("cantrip %s is not in your cantrip list for %s", s, class)
-			return spellList, err
-		}
-	}
-	spellList.Verified = true
-	return spellList, nil
-}
-
-//CheckKnownList func
-func CheckKnownList(spellList *rule.KnownSpellsList) (*rule.KnownSpellsList, error) {
-	numberSpell := spellList.KnownSpells
-	if len(spellList.SpellList) != numberSpell {
-		err := fmt.Errorf("invalid number of known spells")
-		return spellList, err
-	}
-	class := spellList.Class
-	spells := GetFullSpellList(class)
-
-	for _, s := range spellList.SpellList {
-		if !utils.StringInSlice(s, spells) {
-			err := fmt.Errorf("spell added %s is not in your spell list for %s", s, class)
-			return spellList, err
-		}
-		spell := getSpellByName(s)
-		if spell.Level > spellList.SpellMaxLevel {
-			err := fmt.Errorf("spell added %s is higher than your spell maximum of %v level of spell for %s", s, spellList.SpellMaxLevel, class)
-			return spellList, err
-		}
-	}
-	spellList.Verified = true
-	return spellList, nil
-}
-
-//CalcArmorClass func
-func CalcArmorClass(ac *rule.ArmorClass) *rule.ReturnACMessage {
-	// var disvantage string
-	result := new(rule.ReturnACMessage)
-	result.ArmorClassDisvantages = []string{}
-	result.ArmorClassAutomaticallyFails = []string{}
-	result.ArmorClassSpeedReduced = false
-	result.ArmorClassStealthDisvantage = false
-	// armorType, _, armorClassInitial, dexterityMax, strengthMin, stealth, _ := ArmorByName(ac.Armor)
-	armor := ArmorByName(ac.Armor)
-	armorClassInitial := armor.ArmorClass
-	if !utils.StringInSlice(armor.Kind, ac.ArmorProficiency) && ac.Armor != "" {
-		result.ArmorClassDisvantages = []string{"strength", "dexterity"}
-		result.ArmorClassAutomaticallyFails = []string{"spellcast"}
-	}
-	if ac.Ability["strength"] < armor.Strength {
-		result.ArmorClassSpeedReduced = true
-	}
-	var unarmoredDefense int
-
-	if utils.StringInSlice("sorcerous-origin-draconic-resistance", ac.ClassFeatures) && ac.Armor == "" {
-		armorClassInitial = 13
-	}
-	if utils.StringInSlice("unarmored-defense-monk", ac.ClassFeatures) {
-		unarmoredDefense = CalcAbilityModifier(ac.Ability["wisdom"])
-	}
-
-	if utils.StringInSlice("unarmored-defense-barbarian", ac.ClassFeatures) {
-		unarmoredDefense = CalcAbilityModifier(ac.Ability["constitution"])
-	}
-	var fightingStyle int
-	if utils.StringInSlice("fighting-style-defense", ac.ClassFeatures) {
-		fightingStyle = 1
-	}
-
-	abilityModifier := CalcAbilityModifier(ac.Ability["dexterity"])
-	dexterityModifier := abilityModifier
-	if abilityModifier > armor.DexterityModifier {
-		dexterityModifier = armor.DexterityModifier
-	}
-	var shield int
-	if ac.Shield != "" {
-		shield = 2
-
-	}
-	var magic int
-	if ac.ArmorMagicBonus != 0 {
-		magic = ac.ArmorMagicBonus
-		if ac.ShieldMagicBonus != 0 {
-			magic = ac.ArmorMagicBonus + ac.ShieldMagicBonus
-		}
-	}
-	armorClass := armorClassInitial + dexterityModifier + shield + magic + unarmoredDefense + fightingStyle
-	if utils.StringInSlice("unarmored-defense-monk", ac.ClassFeatures) {
-		armorClass = armorClassInitial + dexterityModifier + magic + unarmoredDefense + fightingStyle
-	}
-	if armor.Stealth {
-		result.ArmorClassDisvantages = append(result.ArmorClassDisvantages, "stealth")
-		result.ArmorClassStealthDisvantage = true
-	}
-	result.ArmorClassMaxDexterity = dexterityModifier
-	result.ArmorClass = armorClass
-	return result
-}
-
 //ClassFeatureRoll func
 func ClassFeatureRoll(feature *rule.Feature) (*rule.ReturnCalcMessage, error) {
 	returnResult := new(rule.ReturnCalcMessage)
