+++ usecase/rule/roll_features_usecase_test.go
@@ -1,0 +1,7 @@
+package rule
+
+import "testing"
+
+func TestClassFeatureRoll(t *testing.T) {}
+
+func TestSpecialRaceFeature(t *testing.T) {}
