+++ usecase/rule/roll_monsters_usecase.go
