+++ usecase/rule/roll_monsters_usecase_test.go
