+++ usecase/rule/roll_potion_usecase_test.go
