+++ usecase/rule/roll_power_usecase.go
