+++ usecase/rule/roll_power_usecase_test.go
