+++ usecase/rule/roll_spellcast_usecase.go
