+++ usecase/rule/roll_spellcast_usecase_test.go
