--- usecase/rule/rules_usecase_test.go
@@ -1,22 +1,0 @@
-package rule
-
-import "testing"
-
-func TestListInformation(t *testing.T) {}
-
-func TestFullDescription(t *testing.T) {}
-
-func TestGetConditions(t *testing.T) {}
-
-func TestCheckPreparedSpellsList(t *testing.T) {}
-
-func TestCheckCantripsKnownList(t *testing.T) {}
-
-func TestCheckKnownList(t *testing.T) {}
-
-func TestCalcArmorClass(t *testing.T) {}
-
-func TestClassFeatureRoll(t *testing.T) {}
-
-func TestSpecialRaceFeature(t *testing.T) {}
-
